requirements.txt
pyairtable
Pillow>=10.1.0
face-recognition>=1.3.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
cairosvg>=2.7.1
